Add commments that giving an overall description of  each function, and following by details of each funtions.
Add commments that giving an overall description of what the function will do.

Add description of the first function, which will create a special matrix named "makeCacheMatrix", with a list containing function to set/get the value of the matrix, and its inverse.

The second function is used to compute the inverse of the special matrix returned by "makeCacheMatrix ", with detailed steps to achive the function.